+++ examples/advanced/src/gen/models/ts/petController/UploadFile.ts
@@ -1,4 +1,4 @@
-import type { ApiResponse } from '../ApiResponse'
+import type { ApiResponse } from '../ApiResponse.js'
 
 export type UploadFilePathParams = {
   /**