+++ packages/react/src/shared/utils/squashExportNodes.ts
@@ -21,12 +21,7 @@ export function squashExportNodes(node: DOMElement): Array<KubbFile.Export> {
 
     if (childNode.nodeName === 'kubb-export' && !childNode.attributes.print) {
       const attributes = childNode.attributes as React.ComponentProps<typeof File.Export>
-      exports.push({
-        name: attributes['name'],
-        path: attributes['path'],
-        isTypeOnly: attributes.isTypeOnly,
-        asAlias: attributes['asAlias'],
-      })
+      exports.push(attributes)
     }
   }