+++ packages/swagger-client/src/components/Client.test.tsx
@@ -28,6 +28,7 @@ describe('<Client/>', async () => {
       importPath: '@kubb/swagger-client/client',
     },
     baseURL: '',
+    extName: undefined,
   }
   const plugin = { options } as Plugin<PluginClient>
   const og = await new OperationGenerator(options, {