+++ packages/swagger-zodios/src/plugin.ts
@@ -20,8 +20,10 @@ export const pluginZodios = createPlugin<PluginZodios>((options) => {
   return {
     name: pluginZodiosName,
     options: {
+      extName: output.extName,
       name: trimExtName(output.path),
       baseURL: undefined,
+      includeOperationIdAsAlias: false,
     },
     pre: [pluginOasName, pluginZodName],
     resolvePath(baseName) {
@@ -51,8 +53,10 @@ export const pluginZodios = createPlugin<PluginZodios>((options) => {
 
       const operationGenerator = new OperationGenerator(
         {
+          extName: output.extName,
           name: trimExtName(output.path),
           baseURL,
+          includeOperationIdAsAlias: options.output?.includeOperationIdAsAlias ?? false,
         },
         {
           oas,