+++ packages/swagger-tanstack-query/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @kubb/swagger-tanstack-query
 
+## 2.25.1
+
+### Patch Changes
+
+- [#1110](https://github.com/kubb-labs/kubb/pull/1110) [`3f27fab`](https://github.com/kubb-labs/kubb/commit/3f27fab6ce329d86fd432fc8933890efe07f8319) Thanks [@stijnvanhulle](https://github.com/stijnvanhulle)! - always add extName to import when defined
+
+- Updated dependencies [[`3f27fab`](https://github.com/kubb-labs/kubb/commit/3f27fab6ce329d86fd432fc8933890efe07f8319)]:
+  - @kubb/swagger-client@2.25.1
+  - @kubb/swagger-zod@2.25.1
+  - @kubb/plugin-oas@2.25.1
+  - @kubb/swagger-ts@2.25.1
+  - @kubb/parser-ts@2.25.1
+  - @kubb/react@2.25.1
+  - @kubb/core@2.25.1
+  - @kubb/oas@2.25.1
+  - @kubb/fs@2.25.1
+
+## 2.25.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@2.25.0
+  - @kubb/fs@2.25.0
+  - @kubb/oas@2.25.0
+  - @kubb/parser-ts@2.25.0
+  - @kubb/plugin-oas@2.25.0
+  - @kubb/react@2.25.0
+  - @kubb/swagger-client@2.25.0
+  - @kubb/swagger-ts@2.25.0
+  - @kubb/swagger-zod@2.25.0
+
 ## 2.24.0
 
 ### Patch Changes